:recycle: Separate file for operation
